move proposals page and proposals details page to proposal package
